+++ src/main/java/subway/domain/SubwayManager.java
@@ -1,5 +1,8 @@
 package subway.domain;
 
+import subway.domain.section.SectionRepository;
+import subway.domain.station.Station;
+import subway.domain.station.StationRepository;
 import subway.exception.NoSuchStationException;
 import subway.exception.SameStationException;
 import subway.view.InputView;
