+++ src/main/java/subway/domain/line/LineRepository.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import subway.exception.NoSuchLineException;
