+++ src/main/java/subway/domain/station/StationRepository.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.station;
 
 import subway.exception.NoSuchStationException;
